refact(getTotal, getSubTotal) - BigDecimal for returns monetary values

+++ src/main/java/com/meuportifolio/curso/entities/Product.java
@@ -1,6 +1,7 @@
 package com.meuportifolio.curso.entities;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
@@ -28,7 +29,7 @@ public class Product implements Serializable {
 	private Long id;
 	private String name;
 	private String description;
-	private Double price;
+	private BigDecimal price;
 	private String imgUrl;
 
 	@ManyToMany
@@ -41,7 +42,7 @@ public class Product implements Serializable {
 	public Product() {
 	}
 
-	public Product(Long id, String name, String description, Double price, String imgUrl) {
+	public Product(Long id, String name, String description, BigDecimal price, String imgUrl) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -74,11 +75,11 @@ public class Product implements Serializable {
 		this.description = description;
 	}
 
-	public Double getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
 
-	public void setPrice(Double price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
